+++ src/lib/instance.ts
@@ -6,11 +6,21 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-import { supportedProjects } from "./project";
+import { type supportedProjects } from "./project";
 import {
 	version,
 	repository,
 } from "../../package.json" assert { type: "json" };
+
+// Testing has shown that top 200 fediverse servers run these projects' software.
+// This way, we can avoid checking all 10+ projects we support
+const POPULAR_SOFTWARES = [
+	"mastodon",
+	"fedibird",
+	"misskey",
+	"friendica",
+	"sharkey",
+] satisfies Array<keyof typeof supportedProjects>;
 
 interface Instance {
 	domain: string;
@@ -19,23 +29,44 @@ interface Instance {
 	total_users: number;
 }
 
+const query = `
+	query ($softwarename: String!) {
+		nodes(softwarename: $softwarename, status: "UP") {
+			domain
+			score
+			active_users_monthly
+			total_users
+		}
+	}
+`;
+
+const CACHE_TTL_MS = 24 * 60 * 60 * 1000; // 24 hours
+const instancesCache: Map<
+	keyof typeof supportedProjects,
+	{ value: Instance[] | undefined; lastFetched: number }
+> = new Map();
+
 const getInstancesForProject = async (
 	project: keyof typeof supportedProjects,
 ): Promise<Instance[]> => {
+	if (
+		instancesCache.get(project)?.value &&
+		Date.now() - instancesCache.get(project)!.lastFetched < CACHE_TTL_MS
+	) {
+		return instancesCache.get(project)!.value!;
+	}
+
 	let instances: Instance[];
-
-	const headers = new Headers();
-	headers.set("Accept", "application/graphql-response+json; charset=utf-8");
-	headers.append("Accept", "application/json; charset=utf-8");
-	headers.set("Content-Type", "application/json");
-	headers.set("User-Agent", `Share2Fedi/${version} (+${repository.url})`);
-
 	try {
 		const response = await fetch("https://api.fediverse.observer/", {
-			headers,
+			headers: {
+				Accept: "application/graphql-response+json, application/json",
+				"Content-Type": "application/json",
+				"User-Agent": `Share2Fedi/${version} (+${repository.url})`,
+			},
 			body: JSON.stringify({
-				query: `query($project:String!){nodes(status:"UP",softwarename:$project){domain score active_users_monthly total_users}}`,
-				variables: { project },
+				query,
+				variables: { softwarename: project },
 			}),
 			method: "POST",
 		});
@@ -43,22 +74,30 @@ const getInstancesForProject = async (
 		instances = json.data.nodes;
 	} catch (error) {
 		console.error(`Could not fetch instances for "${project}"`, error);
-		return [];
+		return []; // not caching invalid result
 	}
 
-	return instances.filter(
+	const domains = instances.filter(
 		(instance) =>
 			instance.score > 90 &&
 			// sanity check for some spammy-looking instances
 			instance.total_users >= instance.active_users_monthly,
 	);
+
+	// only cache if not empty
+	if (domains.length > 0) {
+		instancesCache.set(project, {
+			value: domains,
+			lastFetched: Date.now(),
+		});
+	}
+
+	return domains;
 };
 
 export const getPopularInstanceDomains = async (): Promise<string[]> => {
 	const instancesPerProject = await Promise.all(
-		Object.keys(supportedProjects).map((project) =>
-			getInstancesForProject(project),
-		),
+		POPULAR_SOFTWARES.map((project) => getInstancesForProject(project)),
 	);
 	const instances = instancesPerProject.flat();
 	instances.sort((a, b) => b.active_users_monthly - a.active_users_monthly);